backend/requirements.txt
Django>=4.2,<4.3
djangorestframework>=3.14,<3.15
graphene-django>=3.0,<3.1
psycopg2-binary>=2.9,<3.0
requests>=2.25,<3.0
celery>=5.2,<6.0
redis>=4.1,<5.0
django-filter>=2.4,<3.0
django-graphql-jwt>=0.3,<0.4
django-cors-headers>=3.7,<4.0
Faker>=8.1,<9.0
Pillow>=9.0,<12.0
python-dotenv>=0.20,<1.0
